Lit_REVIEW/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-crispy-forms==2.0
sqlparse==0.4.4
